--- a/.changeset/cyan-penguins-smell.md
+++ b/.changeset/cyan-penguins-smell.md
@@ -1,0 +1,5 @@
+---
+'@powersync/react-native': minor
+---
+
+Fixed issue where iOS WebSockets could fail to reconnect after a connection issue.