--- a/packages/common/rollup.config.mjs
+++ b/packages/common/rollup.config.mjs
@@ -2,6 +2,7 @@ import commonjs from '@rollup/plugin-commonjs';
 import inject from '@rollup/plugin-inject';
 import json from '@rollup/plugin-json';
 import nodeResolve from '@rollup/plugin-node-resolve';
+import terser from '@rollup/plugin-terser';
 
 export default (commandLineArgs) => {
   const sourcemap = (commandLineArgs.sourceMap || 'true') == 'true';
@@ -25,8 +26,8 @@ export default (commandLineArgs) => {
         ReadableStream: ['web-streams-polyfill/ponyfill', 'ReadableStream'],
         // Used by can-ndjson-stream
         TextDecoder: ['text-encoding', 'TextDecoder']
-      })
-      // terser()
+      }),
+      terser()
     ],
     // This makes life easier
     external: [