… .changeset/empty-pants-give.md
… .changeset/fuzzy-beers-occur.md
… .changeset/fuzzy-ties-double.md
… packages/common/src/client/sync/stream/AbstractRemote.ts
… packages/common/src/client/sync/stream/AbstractStreamingSyncImplementation.ts
… packages/common/src/client/sync/stream/WebsocketClientTransport.ts
… packages/common/src/db/crud/SyncStatus.ts
… packages/common/src/index.ts
… packages/node/src/db/PowerSyncDatabase.ts
… packages/react-native/rollup.config.mjs
… packages/react-native/src/db/PowerSyncDatabase.ts
… packages/web/src/db/PowerSyncDatabase.ts
… packages/web/src/db/sync/SharedWebStreamingSyncImplementation.ts
… packages/web/src/worker/sync/SharedSyncImplementation.ts
… packages/web/src/worker/sync/SharedSyncImplementation.worker.ts
… packages/web/tests/src/db/AbstractPowerSyncDatabase.test.ts
… packages/web/tests/src/db/PowersyncDatabase.test.ts
… packages/web/tests/stream.test.ts
… packages/web/tests/utils/MockStreamOpenFactory.ts
… packages/web/vitest.config.ts
… tools/diagnostics-app/src/app/views/layout.tsx
… tools/diagnostics-app/src/app/views/sync-diagnostics.tsx
… tools/diagnostics-app/src/library/powersync/ConnectionManager.ts