# Overview

**TL;DR:** Improves connection performance when performing multiple calls to `connect` in rapid succession. 

We currently protect the logic in `connect` calls with an exclusive mutex/lock. This mutex causes invocations to `connect` to be queued.

Certain use-cases such as updating client parameters can result in rapid calls to `connect` when client parameters change. The mutex queuing can cause undesirable delays to this process. Each call to `connect` results in a connection attempt which needs to be awaited before proceeding to the next attempt. 

The work here retains a single connection attempt being active at a time - which prevents possible race conditions. However, **instead of queuing all requests sequentially, the system now buffers and consolidates multiple calls, always using the most recent connection parameters**.

# Implementation

## Key Changes

The connection logic has been moved to a `ConnectionManager` implementation which improves tracking pending asynchronous connect and disconnect operations. 

- **`pendingConnectionOptions`** - Buffers the latest connection parameters
- **`connectingPromise`** - Tracks active connection attempts to prevent races  
- **`disconnectingPromise`** - Coordinates disconnect operations
- **Smart consolidation logic** - Recursive connection checking for queued requests

## Connection Behavior

The timing of `connect` calls greatly affects the outcome of resultant internal `connect` operations. Calling `connect` periodically where the period is longer than the time required to connect will still result in a serial queue of connection attempts. 

Connect calls made with a period less than the time taken to connect or disconnect will be compacted into less requests. 

## Scenario 1: Rapid Successive Calls During Initialization

Calling `.connect` multiple times while the SyncStream implementation is initializating will result in the SyncStream implementation making a single request with the latest parameters. Subsequent calls to `connect`, after the SyncStream has been initialized, will result in a reconnect with the latest (potentially compacted) options. 

```mermaid
sequenceDiagram
    participant Client
    participant PowerSync as PowerSync Client
    participant Buffer as pendingConnectionOptions
    participant SyncStream
    
    Note over Client,SyncStream: SyncStream is initializing asynchronously
    
    Client->>PowerSync: connect(optionsA)
    PowerSync->>Buffer: store optionsA
    PowerSync->>SyncStream: start async initialization
    
    Client->>PowerSync: connect(optionsB)
    PowerSync->>Buffer: update to optionsB
    
    Client->>PowerSync: connect(optionsC)
    PowerSync->>Buffer: update to optionsC
    
    Note over SyncStream: Initialization completes
    SyncStream->>Buffer: get latest options
    Buffer-->>SyncStream: optionsC
    SyncStream->>SyncStream: connect with optionsC
    
    Note over Client,SyncStream: Result: Only optionsC is used, A & B are discarded
```

## Scenario 2: Updates During Active Disconnect

`disconnect` operations are shared internally. Calling `connect` multiple times re-uses a single disconnect operation internally. The latest options are buffered while the disconnect is in progress. 

``` mermaid
sequenceDiagram
    participant Client
    participant PowerSync as PowerSync Client  
    participant Buffer as pendingConnectionOptions
    participant SyncStream as Active SyncStream
    
    Note over SyncStream: SyncStream is connected and running
    
    Client->>PowerSync: connect(newOptionsA)
    PowerSync->>SyncStream: start disconnect()
    PowerSync->>Buffer: store newOptionsA
    
    Client->>PowerSync: connect(newOptionsB)
    PowerSync->>Buffer: update to newOptionsB
    
    Client->>PowerSync: connect(newOptionsC)  
    PowerSync->>Buffer: update to newOptionsC
    
    Note over SyncStream: Disconnect completes
    SyncStream-->>PowerSync: disconnected
    
    PowerSync->>Buffer: get latest options
    Buffer-->>PowerSync: newOptionsC
    PowerSync->>SyncStream: initialize & connect with newOptionsC
    
    Note over Client,SyncStream: Result: Seamless transition to latest options
```

## Edge Cases Handled

- **Concurrent multi-tab scenarios** - The initialization of the shared sync implementation is guarded by navigator locks in the web implementation of `runExclusive`
- **Connect during disconnect** - Proper serialization with promise coordination
- **Rapid parameter updates** - Latest-wins semantics ensure current user intent

# Performance Comparison

## Test Results: 21 Rapid `connect()` Calls

| Metric | Before (main) | After (this PR) | Improvement |
|--------|---------------|-----------------|-------------|
| **Total Time** | 2.5 seconds | 0.4 seconds | **6.25x faster** |
| **Sync Implementations Created** | 21 | 3 | **7x fewer resources** |
| **Connection Attempts** | 21 sequential | 3 consolidated | **Optimal efficiency** |

### Before (main branch)
![image](https://github.com/user-attachments/assets/57eeb42a-f5e7-4fcc-b4c2-c14dce4af728)
*21 sync stream implementations created, each attempting connection before disposal*

### After (this PR)  
![image](https://github.com/user-attachments/assets/2ab443ac-643c-43e9-b497-5c791f4cbadd)
*Only 3 sync implementations created with intelligent parameter consolidation*

# Breaking Changes
None - this is a pure performance and efficiency improvement with identical public API behavior.

# Testing

This can be tested by using the following development packages.

```
@powersync/common@0.0.0-dev-20250529141956
@powersync/node@0.0.0-dev-20250529141956
@powersync/op-sqlite@0.0.0-dev-20250529141956
@powersync/react-native@0.0.0-dev-20250529141956
@powersync/tanstack-react-query@0.0.0-dev-20250529141956
@powersync/web@0.0.0-dev-20250529141956
```