@obezzad 2025-05-27
Thanks a lot @stevensJourney. Very exciting work! Just tested the dev packages and wanted to share some preliminary observations: after expanding nodes, it seems that one stream keeps reconnecting every few seconds without any action from my end. Is this expected?

https://github.com/user-attachments/assets/2860ed8a-a822-4588-8159-334506fc3c28